Bugfix: Eliminate double-click of "Reply" buttons to reply input in another comment
Logic originally set up such that when attempting to open a different reply input after a previous, there would be a one-click delay in the input opening. This was due to the TOGGLE_REPLY_INPUT action using a simple NOT operator. Extra logic was needed to only close the reply input if the reply button that was clicked was the same reply button as before. Otherwise, if another reply button was clicked, just switch the idOfCommentReceivingReply state and keep the toggle open'